+++ src/annbatch/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from importlib.metadata import version
 
-from . import abc, samplers, types
+from . import samplers, types
 from .io import DatasetCollection, write_sharded
 from .loader import Loader
 from .samplers import ChunkSampler
@@ -15,6 +15,5 @@ __all__ = [
     "DatasetCollection",
     "types",
     "write_sharded",
-    "abc",
     "samplers",
 ]
