+++ src/annbatch/samplers/__init__.py
@@ -1,3 +1,4 @@
+from . import abc
 from ._chunk_sampler import ChunkSampler, DistributedRandomSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler
@@ -7,4 +8,5 @@ __all__ = [
     "DistributedRandomSampler",
     "RandomSampler",
     "SequentialSampler",
+    "abc",
 ]
