+++ tests/test_dataset.py
@@ -14,8 +14,8 @@ import scipy.sparse as sp
 import zarr
 
 from annbatch import Loader, write_sharded
-from annbatch.abc import Sampler
 from annbatch.samplers import SequentialSampler
+from annbatch.samplers.abc import Sampler
 
 try:
     from cupy import ndarray as CupyArray
